Start factoring out a BaseSQLQueryEngine
We do this by renaming the existing MssqlQueryEngine to
BaseSQLQueryEngine (without modifying it at all) and then making
MssqlQueryEngine be an empty subclass. We can then move over the
MSSQL-specific elements and leave behind the common SQL code.
