+++ tests/query_engines/test_mssql.py
@@ -6,22 +6,6 @@ import sqlalchemy
 from cohortextractor.query_engines import mssql
 
 
-@pytest.mark.parametrize(
-    "lst,size,expected",
-    [
-        ([], 10, []),
-        (range(7), 3, [[0, 1, 2], [3, 4, 5], [6]]),
-        (range(4), 6, [[0, 1, 2, 3]]),
-        (range(12), 4, [[0, 1, 2, 3], [4, 5, 6, 7], [8, 9, 10, 11]]),
-    ],
-)
-def test_split_list_into_batches(lst, size, expected):
-    lst = list(lst)
-    results = mssql.split_list_into_batches(lst, size)
-    results = list(results)
-    assert results == expected
-
-
 def test_mssql_date_types():
     # Note: it would be nice to parameterize this test, but given that the
     # inputs are SQLAlchemy expressions I don't know how to do this without
